rotkehlchen: Remove lact
There were some freezes that could only be resolved by a hard reset. I
am suspecting that gpu undervolting or the amdgpu ppfeaturemask might
have something to do with that because there were some amdgpu-related
errors in the journal, e.g.

[drm:amdgpu_job_submit [amdgpu]] *ERROR* Trying to push to a killed entity

However, the freeze seems to occur especially after a near-oom, but that
might be a coincidence. MemTest86 says that the memory is good.

So anyways, maybe removing lact will help, maybe not. No idea.